Modificación para que un usuario premium no pueda agregar sus mismos productos al carrito, Agrego el owner al producto Evito que se borre prosucto de otro owner salvo que sea admin

+++ src/controllers/controller.products.js
@@ -51,45 +51,47 @@ router.put('/:pid', authToken, adminAccess, async (req, res, next) => {
     try {
         const product = await Products.getProductById(pid);
         if (product) {
-            const { title, measurement, thumbnails, stock, price, description, category } = req.body;
-            if (!isNaN(price) & (parseFloat(price) < 0))
-            {
-                CustomError.createError({
-                    name: "Error actualizando el producto",
-                    cause: "El precio debe ser mayor que 0",
-                    message: "Error al intentar actualizar el producto.",
-                    code: EnumErrors.INVALID_PARAM
-                });
-            }
-            const productInfo = {
-                title,
-                measurement,
-                thumbnails,
-                stock,
-                price,
-                description,
-                category
-            };
-            const result = await Products.updateProduct(pid, productInfo);
-            if (result) {
-                res.status(200).json({ message: 'Producto actualizado satisfactoriamente' });
-            }
-            else {
-                res.status(500).json({ error: 'Internal server error' });
+            if ((product.owner == req.user.email) || (req.user.role = 'admin')) {
+                const { title, measurement, thumbnails, stock, price, description, category } = req.body;
+                if (!isNaN(price) & (parseFloat(price) < 0)) {
+                    CustomError.createError({
+                        name: "Error actualizando el producto",
+                        cause: "El precio debe ser mayor que 0",
+                        message: "Error al intentar actualizar el producto.",
+                        code: EnumErrors.INVALID_PARAM
+                    });
+                }
+                const productInfo = {
+                    title,
+                    measurement,
+                    thumbnails,
+                    stock,
+                    price,
+                    description,
+                    category
+                };
+                const result = await Products.updateProduct(pid, productInfo);
+                if (result) {
+                    res.status(200).json({ message: 'Producto actualizado satisfactoriamente' });
+                }
+                else {
+                    res.status(500).json({ error: 'Internal server error' });
+                }
+            } else {
+                res.status(400).json({ status: 'error', message: 'No tiene permisos para eliminar este producto' });
             }
         } else {
             res.status(404).json({ error: 'Producto no encontrado' });
         }
     } catch (error) {
-        if (error.code == EnumErrors.INVALID_PARAM)
-        {
+        if (error.code == EnumErrors.INVALID_PARAM) {
             next(error);
         }
         else {
             req.logger.error(error.message);
             res.status(500).json({ error: 'Internal server error.' });
         }
-        
+
     }
 });
 
@@ -104,7 +106,8 @@ router.post('/', authToken, adminAccess, uploader.single('file'), async (req, re
             thumbnails: req.file.filename,
             stock,
             price,
-            description, category
+            description, category,
+            owner: req.user.email
         };
         const newProduct = await Products.addProduct(productInfo);
         res.status(201).json({ message: newProduct });
@@ -123,15 +126,20 @@ router.delete('/:pid', authToken, adminAccess, async (req, res) => {
     const { pid } = req.params;
     const productToDel = await Products.getProductById(pid);
     if (productToDel) {
-        const result = await Products.deleteProduct(pid);
-        if (result) {
-            res.status(200).json({ status: 'success', message: 'Producto eliminado satisfactoriamente !!!' });
+        if ((productToDel.owner == req.user.email) || (req.user.role = 'admin')) {
+            const result = await Products.deleteProduct(pid);
+            if (result) {
+                res.status(200).json({ status: 'success', message: 'Producto eliminado satisfactoriamente !!!' });
+            } else {
+                res.status(500).json({ status: 'error', message: 'Internal server error' });
+            }
         } else {
-            res.status(500).json({ status: 'error', message: 'Internal server error' });
+            res.status(400).json({ status: 'error', message: 'No tiene permisos para eliminar este producto' });
         }
     } else {
         res.status(404).json({ status: 'error', message: 'Producto no encontrado' });
     }
+
 });
 
 router.post('/insertMany', authToken, adminAccess, async (req, res) => {
@@ -141,7 +149,7 @@ router.post('/insertMany', authToken, adminAccess, async (req, res) => {
         await Products.addManyProducts(jsonProducts);
         res.status(200).json({ status: 'success', message: 'Productos agregados!!!' });
     } catch (error) {
-        res.status(500).json({ status: 'error', message: error.message});
+        res.status(500).json({ status: 'error', message: error.message });
     }
 });
